직급, 역할 조회를 layout에서.
