+++ src/components/home/HomePage.tsx
@@ -1,61 +1,6 @@
 import { Navigate } from "react-router-dom";
-import requestDeprecated, {
-  RequestFailHandler,
-  RequestSuccessHandler,
-} from "../../lib/requestDeprecated";
-import { useRecoilState } from "recoil";
-import hierarchyListState, {
-  Hierarchy,
-} from "../../states/data/hierarchyListState";
-import roleListState, { Role } from "../../states/data/roleListState";
-import { useEffect } from "react";
 
 function HomePage() {
-  const [, setHierarchyList] = useRecoilState(hierarchyListState);
-  const [, setRoleList] = useRecoilState(roleListState);
-
-  //set enum table
-  //직급 리스트
-  const hierarchyRequest = () => {
-    const handelRequestSuccess: RequestSuccessHandler = (res) => {
-      const data: Hierarchy[] = res.data;
-      setHierarchyList(data);
-    };
-    const handelRequestFail: RequestFailHandler = (e) => {
-      alert((e.response.data as { code: string; message: string }).message);
-    };
-
-    requestDeprecated("GET", `/hierarchy`, {
-      withToken: false,
-      useMock: false,
-      onSuccess: handelRequestSuccess,
-      onFail: handelRequestFail,
-    });
-  };
-
-  //권한 리스트
-  const roleRequest = () => {
-    const handelRequestSuccess: RequestSuccessHandler = (res) => {
-      const data: Role[] = res.data;
-      setRoleList(data);
-    };
-    const handelRequestFail: RequestFailHandler = (e) => {
-      alert((e.response.data as { code: string; message: string }).message);
-    };
-
-    requestDeprecated("GET", `/role`, {
-      withToken: false,
-      useMock: false,
-      onSuccess: handelRequestSuccess,
-      onFail: handelRequestFail,
-    });
-  };
-
-  useEffect(() => {
-    hierarchyRequest();
-    roleRequest();
-  }, []);
-
   return <Navigate to="clients" />;
 }
